Conflicts:
	Viewer/Bugfix/views.py
	Viewer/static/list.js
	Viewer/templates/home.html
	scripts/monitors/monitor_count.py
	scripts/wbxmldecoder.py

+++ scripts/monitors/monitor_log.py
@@ -195,15 +195,16 @@ class MonitorLog(Monitor):
                             event[k] = client[k]
 
                 print >>raw_log, ef.format(event).encode('utf-8')
+
         zipped_log_path = raw_log_prefix + '.zip'
         zipped_file = zipfile.ZipFile(zipped_log_path, 'w', zipfile.ZIP_DEFLATED)
         zipped_file.write(raw_log_path)
         os.unlink(raw_log_path)
-
-        for trace in self.traces:
-            trace_path = trace.write_file()
-            zipped_file.write(trace_path)
-            os.unlink(trace_path)
+        if self.traces:
+            for trace in self.traces:
+                trace_path = trace.write_file()
+                zipped_file.write(trace_path)
+                os.unlink(trace_path)
         zipped_file.close()
         return zipped_log_path
 
@@ -215,7 +216,6 @@ class MonitorErrors(MonitorLog):
         kwargs.setdefault('msg', 'Error count')
         kwargs.setdefault('rate_msg', 'ERROR rate')
         MonitorLog.__init__(self, *args, **kwargs)
-        self.trace_enabled = True
 
 
 class MonitorWarnings(MonitorLog):
